requirements.txt
django >= 2.0, < 2.1
django-cors-headers >= 2.1, < 2.2
djangorestframework >= 3.7, < 3.8
